+++ docs/tasks/artifacts/sv_corpus_grad/class_qualifier_extraction_mangle/control_foo_qualifier_as_prefix.sv
@@ -1,0 +1,4 @@
+module top;
+  int y, m;
+  initial y = foo_qualifier m;
+endmodule