+++ docs/tasks/artifacts/sv_corpus_grad/class_qualifier_extraction_mangle/control_foo_qualifier_assign.sv
@@ -1,0 +1,4 @@
+module top;
+  int y;
+  initial y = null foo_qualifier:=;
+endmodule